+++ server-application/src/main/java/com/genlab/serverapplication/services/ctservice/Polyhybrid/CTPolyHybrid.java
@@ -4,8 +4,8 @@ import com.genlab.serverapplication.models.CTResult;
 
 public interface CTPolyHybrid {
 
-    CTResult polyhybrid (int n, int h, int d, int r, int D, int R, int T);
+    CTResult polyhybrid (double n, double h, double d, double r, double D, double R, double T);
 
-    CTResult multiplealleles (int locus1, int locus2, int locus3, int locus4, int locus5);
+    CTResult multiplealleles (double locus1, double locus2, double locus3, double locus4, double locus5);
 
 }
